cards equal width, visible sections titles
